{{cookiecutter.project_slug}}/requirements/testing.txt
-r base.txt

black==19.3b0
coverage==4.5.3
django-extensions==2.2.1
flake8==3.7.8
isort==4.3.21
pipdeptree==0.13.2
factory-boy==2.12.0
unittest-xml-reporting==2.5.1
